scripts/pyproject.toml
[project]
name = "patti-scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "pydub>=0.25.1",
    "scipy>=1.12.0",
]
